% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_functions.R
\name{auth}
\alias{auth}
\title{Get a Dropbox Access Token}
\usage{
auth(envpath = ".env")
}
\arguments{
\item{envpath}{The path to a local .env file}
}
\value{
A short-lived access token.
}
\description{
Internal helper function to manage OAuth2 flow.
It checks for an existing REFRESH_TOKEN in the .env file.
If found, it uses it to get a new access token.
If not found, it initiates the manual "code" flow, gets a
refresh token, and appends it to the .env file.
}
\examples{
\dontrun{
dbx <- auth("/path/to/.env")
}
}
