% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_functions.R
\name{compute_local_hash}
\alias{compute_local_hash}
\title{Compute the Dropbox Content Hash for a Local File}
\usage{
compute_local_hash(local_path)
}
\arguments{
\item{local_path}{Path to the local file.}
}
\value{
A string (hex digest) representing the Dropbox content hash.
}
\description{
Dropbox computes a content hash by:
\enumerate{
\item Splitting the file into 4MB chunks.
\item Calculating the SHA256 hash of each chunk.
\item Concatenating all chunk hashes.
\item Calculating the SHA256 hash of the concatenated string.
Para arquivos muito grandes, talvez seja muito custoso calcular o hash sempre a depender do computador
}
}
