% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_functions.R
\name{download_dropbox_file}
\alias{download_dropbox_file}
\title{Download a File from Dropbox}
\usage{
download_dropbox_file(dbx_token, dropbox_path, local_path)
}
\arguments{
\item{dbx_token}{The short-lived access token}

\item{dropbox_path}{The full path to the file in your Dropbox (e.g., "/Apps/MyApp/data.csv").}

\item{local_path}{The local file path to save the file to (e.g., "./my_data.csv").}
}
\value{
NULL. The file is saved to disk.
}
\description{
Downloads a file from a specified Dropbox path to a local path.
}
\examples{
\dontrun{
library(GEESCHelpersR)

local_path <- "/local/path.csv"
drop_path <- "/dropbox/path.csv"
dbx <- auth("/path/to/.env")

dbx |>
  download_dropbox_file(drop_path, local_path)
}
}
