% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_functions.R
\name{download_dropbox_memory}
\alias{download_dropbox_memory}
\title{Download a File from Dropbox to memory}
\usage{
download_dropbox_memory(dbx_token, dropbox_path)
}
\arguments{
\item{dbx_token}{The short-lived access token}

\item{dropbox_path}{The full path to the file in your Dropbox (e.g., "/Apps/MyApp/data.csv").}
}
\value{
A tempfile read from dropbox
}
\description{
Read file into memory without saving to disk
}
\examples{
\dontrun{
library(GEESCHelpersR)

drop_path <- "/dropbox/path.csv"
dbx <- auth("/path/to/.env")

dbx |>
  download_dropbox_memory(drop_path) |>
  # function to read the file
  read.csv()
}
}
