% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_functions.R
\name{get_dropbox_hash}
\alias{get_dropbox_hash}
\title{Get Dropbox File Content Hash}
\usage{
get_dropbox_hash(dbx_token, dropbox_path)
}
\arguments{
\item{dbx_token}{The short-lived access token}

\item{dropbox_path}{The full path to the file in your Dropbox (e.g., "/Apps/MyApp/data.csv").}
}
\value{
A string containing the 'content_hash' of the file.
}
\description{
Retrieves the metadata for a file in Dropbox, including the 'content_hash'.
}
