% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_functions.R
\name{try_download}
\alias{try_download}
\title{Compare local file with dropbox's, and then load it into memory}
\usage{
try_download(dbx_token, dropbox_path, local_path)
}
\arguments{
\item{dbx_token}{A short-lived access token}

\item{dropbox_path}{The full path to the file in your Dropbox (e.g., "/Apps/MyApp/data.csv").}

\item{local_path}{The local file path to save the file to (e.g., "./my_data.csv").}
}
\value{
The file read either locally or from dropbox. The file is also saved to disk.
}
\description{
Downloads a file from a specified Dropbox path to a local path.
}
\examples{
\dontrun{
library(GEESCHelpersR)

local_path <- "/local/path.csv"
drop_path <- "/dropbox/path.csv"
dbx <- auth("/path/to/.env")

file <- dbx |>
  try_download(drop_path, local_path) |>
  # function to read the file
  read.csv()
}
}
