% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_functions.R
\name{try_sf_download}
\alias{try_sf_download}
\title{Compare local zip (.shp) file with dropbox's, and then load it into memory}
\usage{
try_sf_download(dbx_token, dropbox_path, local_path)
}
\arguments{
\item{dbx_token}{A short-lived access token}

\item{dropbox_path}{The full path to the file in your Dropbox (e.g., "/Apps/MyApp/data.csv").}

\item{local_path}{The local file path to save the file to (e.g., "./my_data.csv").}
}
\value{
the zip tempdir to be read
}
\description{
Loads a zip file into memory using \link[GEESCHelpersR]{try_download}
}
\examples{
\dontrun{
library(GEESCHelpersR)
library(sf)

local_path <- "/local/path.zip"
drop_path <- "/dropbox/path.zip"
dbx <- auth("/path/to/.env")

file <- dbx |>
  try_sf_download(drop_path, local_path) |>
  # function to read the file
  read_sf()
}
}
